+++ tests/provider/dwd/radar/test_api_most_recent.py
@@ -2,6 +2,7 @@
 # Copyright (C) 2018-2021, earthobservations developers.
 # Distributed under the MIT License. See LICENSE for more info.
 import pytest
+from dirty_equals import IsDict, IsList, IsStr
 
 from wetterdienst.provider.dwd.radar import (
     DwdRadarParameter,
@@ -113,7 +114,8 @@ def test_radar_request_site_most_recent_sweep_vol_v_hdf5():
     assert hdf["/dataset1/how"].attrs.get("scan_index") == 2
 
 
-def test_radar_request_radolan_cdc_most_recent():
+@pytest.mark.remote
+def test_radar_request_radolan_cdc_most_recent(radar_locations):
     """
     Example for testing radar sites most recent RADOLAN_CDC.
     """
@@ -139,50 +141,22 @@ def test_radar_request_radolan_cdc_most_recent():
     requested_attrs = wrl.io.parse_dwd_composite_header(requested_header)
 
     # Verify data.
-    attrs = {
-        "producttype": "SF",
-        "datetime": request.start_date.to_pydatetime(),
-        "precision": 0.1,
-        "formatversion": 3,
-        "intervalseconds": 86400,
-        "nrow": 900,
-        "ncol": 900,
-        "radarlocations": [
-            "asb",
-            "boo",
-            "ros",
-            "hnr",
-            "umd",
-            "pro",
-            "ess",
-            "fld",
-            "drs",
-            "neu",
-            "nhb",
-            "oft",
-            "eis",
-            "tur",
-            "isn",
-            "fbg",
-            "mem",
-        ],
-    }
-
-    # radar locations can change over time -> check if at least 10 radar locations
-    # were found and at least 5 of them match with the provided one
-    assert len(requested_attrs["radarlocations"]) >= 10
-    assert len(list(set(requested_attrs["radarlocations"]) & set(attrs["radarlocations"]))) >= 5
-
-    skip_attrs = [
-        "radolanversion",
-        "radardays",
-        "radarlocations",
-        "radarid",
-        "maxrange",
-        "datasize",
-    ]
-    for attr in skip_attrs:
-        requested_attrs.pop(attr, None)
-    del attrs["radarlocations"]
+    attrs = IsDict(
+        {
+            "datasize": 1620000,
+            "datetime": request.start_date.to_pydatetime(),
+            "formatversion": 3,
+            "intervalseconds": 86400,
+            "maxrange": "150 km",
+            "ncol": 900,
+            "nrow": 900,
+            "precision": 0.1,
+            "producttype": "SF",
+            "radardays": IsList(IsStr(regex=" [0-9]?[0-9]?|".join(radar_locations)), length=(10, len(radar_locations))),
+            "radarid": "10000",
+            "radarlocations": IsList(IsStr(regex="|".join(radar_locations)), length=(10, len(radar_locations))),
+            "radolanversion": "2.29.1",
+        }
+    )
 
     assert requested_attrs == attrs