+++ CHANGELOG.rst
@@ -5,6 +5,7 @@ Development
 ***********
 
 - Update wetterdienst explorer with clickable stations and slighly changed layout
+- Improve radar tests and certain dict comparisons
 
 0.50.0 (03.12.2022)
 *******************